requirements.txt
livekit-agents>=0.12.9
livekit-plugins-deepgram>=0.6.16
livekit-plugins-openai[vertex]>=0.10.10
livekit-plugins-silero>=0.7.4
livekit-plugins-rag>=0.2.3
livekit-plugins-azure>=0.5.3
python-dotenv~=1.0
aiofile~=3.8.8
